// modules/blocks-subgraph/generated/blocks-subgraph-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    BigDecimal: string;
    BigInt: string;
    Bytes: string;
};

export type Block = {
    __typename?: 'Block';
    author?: Maybe<Scalars['String']>;
    difficulty?: Maybe<Scalars['BigInt']>;
    gasLimit?: Maybe<Scalars['BigInt']>;
    gasUsed?: Maybe<Scalars['BigInt']>;
    id: Scalars['ID'];
    number: Scalars['BigInt'];
    parentHash?: Maybe<Scalars['String']>;
    receiptsRoot?: Maybe<Scalars['String']>;
    size?: Maybe<Scalars['BigInt']>;
    stateRoot?: Maybe<Scalars['String']>;
    timestamp: Scalars['BigInt'];
    totalDifficulty?: Maybe<Scalars['BigInt']>;
    transactionsRoot?: Maybe<Scalars['String']>;
    unclesHash?: Maybe<Scalars['String']>;
};
// ...
export type Block_Height = {
    hash?: InputMaybe<Scalars['Bytes']>;
    number?: InputMaybe<Scalars['Int']>;
    number_gte?: InputMaybe<Scalars['Int']>;
};

export enum Block_OrderBy {
    Author = 'author',
    Difficulty = 'difficulty',
    GasLimit = 'gasLimit',
    GasUsed = 'gasUsed',
    Id = 'id',
    Number = 'number',
    ParentHash = 'parentHash',
    ReceiptsRoot = 'receiptsRoot',
    Size = 'size',
    StateRoot = 'stateRoot',
    Timestamp = 'timestamp',
    TotalDifficulty = 'totalDifficulty',
    TransactionsRoot = 'transactionsRoot',
    UnclesHash = 'unclesHash',
}

export enum OrderDirection {
    Asc = 'asc',
    Desc = 'desc',
}

export type Query = {
    __typename?: 'Query';
    /** Access to subgraph metadata */
    _meta?: Maybe<_Meta_>;
    block?: Maybe<Block>;
    blocks: Array<Block>;
};
// ...
export type _Meta_ = {
    __typename?: '_Meta_';
    /**
     * Information about a specific subgraph block. The hash of the block
     * will be null if the _meta field has a block constraint that asks for
     * a block number. It will be filled if the _meta field has no block constraint
     * and therefore asks for the latest  block
     *
     */
    block: _Block_;
    /** The deployment ID */
    deployment: Scalars['String'];
    /** If `true`, the subgraph encountered indexing errors at some past block */
    hasIndexingErrors: Scalars['Boolean'];
};

export enum _SubgraphErrorPolicy_ {
    /** Data will be returned even if the subgraph has indexing errors */
    Allow = 'allow',
    /** If the subgraph has indexing errors, data will be omitted. The default. */
    Deny = 'deny',
}

export type BlocksQueryVariables = Exact<{
    skip?: InputMaybe<Scalars['Int']>;
    first?: InputMaybe<Scalars['Int']>;
    orderBy?: InputMaybe<Block_OrderBy>;
    orderDirection?: InputMaybe<OrderDirection>;
    where?: InputMaybe<Block_Filter>;
    block?: InputMaybe<Block_Height>;
}>;

export type BlocksQuery = {
    __typename?: 'Query';
    blocks: Array<{ __typename?: 'Block'; id: string; number: string; timestamp: string }>;
};

export type BlockFragment = { __typename?: 'Block'; id: string; number: string; timestamp: string };